requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
google-cloud-bigquery==3.30.0
openpyxl==3.1.2
pandas==2.2.2
Werkzeug==3.0.1
gunicorn==21.2.0
db-dtypes==1.2.0
google-cloud-secret-manager
